requirements.txt
tablib==3.1.0
django-tables2==2.4.1
django-crispy-forms==1.13.0
openpyxl==3.0.9